MonsterBot/requirements.txt
ccxt
pandas
pandas_ta
python-dotenv
pyyaml
optuna
termcolor
python-telegram-bot
streamlit
plotly
websocket-client
